Trend reports were moved to separate files.
Closes #63.
